requirements.txt
typer[all]
notion-client
openai
python-dotenv
rich
pypdf
